fix: folder restructure
