+++ src/app/(discover)/searchDetails/(search)/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+import { useState } from "react";
+import Card from "../../(components)/Card"
+
+export default function page() {
+  return (
+    <div>
+      <div className="flex gap-4">
+        <Card />
+      </div>
+    </div>
+  );
+}
