+++ src/app/topics/page.tsx
@@ -1,0 +1,3 @@
+export default  function Topic(){
+    return <h1 className="text-red-600 text-4xl text-center mt-44">This page is Topics</h1>
+}
