+++ tailwind.config.ts
@@ -4,7 +4,6 @@ module.exports = {
     "./app/**/*.{js,ts,jsx,tsx,mdx}",
     "./pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./components/**/*.{js,ts,jsx,tsx,mdx}",
-
     "./src/**/*.{js,ts,jsx,tsx,mdx}",
   ],
   theme: {
@@ -16,6 +15,7 @@ module.exports = {
       },
       colors: {
         'custom-bg': '#191A1F', 
+        'custom-them': '#F65930', 
       },
     },
   },
